Added expand/collapse functionality for iframe usage
With three new options in config.js

expandable = If the graph tool appears in an iframe, enabling this displays a new button that allows the iframe to take over the host page.

expandableOnly = Restricts usage of the graph tool until after the user has expanded it.

headerHeight = If the site owner wants to enabled expanded mode but wants the header to remain visible, specify height of header here.